Everything should be resolved Still working on the close and shutdown of the server that make some problems somehow #4
