+++ src/server/Main.java
@@ -2,28 +2,20 @@ package server;
 
 import java.util.Scanner;
 
+
 public class Main {
 
     public static final int PORT = 6123;
 
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
-        String input;
+        String input = "";
 
         MyServer server = new MyServer(PORT, new PlayerHandler(), 3);
         server.start();
 
-        // Wait for some time to allow server to finish
-        try {
-            Thread.sleep(2000);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
 
-
-
-
-        while (true) {
+        while (!input.equals("stop")) {
             input = scanner.nextLine();
 
             if (input.equalsIgnoreCase("stop")) {
@@ -37,6 +29,8 @@ public class Main {
                 System.exit(0);
             }
         }
+
+
 
 
     }
