+++ test/java/server/BoardTest.java
@@ -4,13 +4,10 @@ package server;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import server.Tile.Bag;
-import static server.PlayerHandler.client1;
-
-import java.util.Scanner;
 
 public class BoardTest {
 
-    public static final int PORT = 6123;
+
     final String DICTIONARY_1 = "test/resources/dictionary_1.txt";
     final String DICTIONARY_2 = "test/resources/dictionary_2.txt";
 
@@ -44,51 +41,5 @@ public class BoardTest {
         Assertions.assertTrue(board.dictionaryLegal(word));
     }
 
-    @Test
-    public void testServer(){
-        Scanner scanner = new Scanner(System.in);
-        String input;
-
-        MyServer server = new MyServer(PORT, new PlayerHandler(), 3);
-        server.start();
-
-        // Wait for some time to allow server to finish
-        try {
-            Thread.sleep(2000);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-
-        //Test of the server
-
-        // Create and start clients
-        client1(PORT);
-        client1(PORT);
-        client1(PORT);
-
-        client1(PORT);
-
-        client1(PORT);
-        client1(PORT);
-        client1(PORT);
-
-
-        while (true) {
-            input = scanner.nextLine();
-            if (input.equalsIgnoreCase("stop")) {
-                System.out.println("Stopping the main program...");
-
-                // Perform any necessary cleanup or shutdown tasks
-                server.close();
-
-                // Close resources and exit the program
-                scanner.close();
-                System.exit(0);
-            }
-        }
-
-
-
-    }
 
 }
